Cambios Issues

+++ 04_area_triangulo/Ejercicio4/src/com/company/Main.java
@@ -12,10 +12,10 @@ public class Main {
         BufferedReader br =  new BufferedReader( new InputStreamReader(System.in));
 
         System.out.println("Introduce la base: ");
-        int base = Integer.parseInt(br.readLine());
+        double base = Integer.parseInt(br.readLine());
 
         System.out.println("Introduce la altura: ");
-        int altura = Integer.parseInt(br.readLine());
+        double altura = Integer.parseInt(br.readLine());
 
         double area = 2/(base*altura);
